+++ packages/cli/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 ## Recent Changes
 
-- Enhancement: Exposed new option `modifyVersion` for the `zowe zos-jobs delete job` command. [#1092](https://github.com/zowe/zowe-cli/issues/1092)
+- Enhancement: Exposed new option `modifyVersion` for the `zowe zos-jobs delete job` and `zowe zos-jobs cancel job` commands. [#1092](https://github.com/zowe/zowe-cli/issues/1092)
 
 ## `6.33.1`